requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==45.0.3
Django==5.2.1
django-celery-beat==2.8.1
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-ratelimit==4.1.0
django-redis==6.0.0
django-restframework==0.0.1
django-silk==5.4.0
django-timezone-field==7.1
django-use-email-as-username==1.4.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
flower==2.0.1
gprof2dot==2025.4.14
gTTS==2.5.4
hiredis==3.2.1
httplib2==0.22.0
humanize==4.12.3
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
oauth2client==4.1.3
openapi-codec==1.3.2
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
requests==2.32.3
rsa==4.9.1
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
